git-svn-id: http://192.168.1.50/svn/bhp_dispatch@151 f4864c45-e38a-4eae-8a0a-e093f1d4e12c

+++ classes/base_prepare_device.py
@@ -4,7 +4,7 @@ from uuid import uuid4
 from datetime import datetime
 from tastypie.models import ApiKey
 from django.db import IntegrityError
-from django.db.models import Model, get_model, get_models, get_app, Max
+from django.db.models import Model, get_model, get_models, get_app, Max, ForeignKey, OneToOneField
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.auth.models import User, Group, Permission
 from bhp_base_model.classes import BaseListModel
@@ -91,11 +91,42 @@ class BasePrepareDevice(Base):
             self.dispatch_model_as_json(list_model)
 
     def reset_model(self, model_cls):
+        """Deletes all instances of the given model and its audit log entries 
+        
+        """
+        if not model_cls:
+            raise TypeError('Please provide a model to delete.')
         using = self.get_using_destination()
+#        print '    deleting {0}...'.format(model_cls._meta.object_name)
+#        try:
+#            self.reset_model_raw_sql(model_cls)
+#            if "history" in dir(model_cls):
+#                if model_cls.history.using(using).all().count() > 0:
+#                    self.delete_audit_instances(model_cls)
+#
+#        except IntegrityError as e:
+#            logger.info(e)
+#            if 'Cannot delete or update a parent row' in e.args[1]:
+#                if '_audit' in e.args[1]:
+#                    # assume Integrity error was because of an undeleted related Audit model
+#                    self.delete_audit_instances(model_cls)
+#                    self.reset_model_raw_sql(model_cls)
+#                else:
+#                    self.delete_depended_model_instances(model_cls)
+##                    self.reset_model_raw_sql(model_cls)
+#                    if "history" in dir(model_cls):
+#                        if model_cls.history.using(using).all().count() > 0:
+#                            self.delete_audit_instances(model_cls)
         for instance in model_cls.objects.using(using).all():
             print '    deleting {0}...'.format(instance)
             try:
-                instance.delete(using=using)
+                #Delete all the audit logs for the model
+                #self.reset_model(model_cls)
+                if "history" in dir(model_cls):
+                    if model_cls.history.using(using).all().count() > 0:
+                        self.delete_audit_instances(model_cls)
+                    instance.delete(using=using)
+                    print '    {0} -> {1} deleted..'.format(instance._meta.object_name, instance)
             except IntegrityError as e:
                 logger.info(e)
                 if 'Duplicate' in e.args[1]:
@@ -105,25 +136,77 @@ class BasePrepareDevice(Base):
                         # assume Integrity error was because of an undeleted related Audit model
                         self.delete_audit_instances(model_cls)
                         instance.delete(using=using)
-                        
+                    else:
+                        self.delete_depended_model_instances(model_cls)
+                        instance.delete(using=using)
+                elif 'Cannot add or update a child row' in e.args[1]:
+                    if '_audit' in e.args[1]: 
+                        raise
                     else:
                         raise
                 else:
                     raise
             except:
                     raise
-            print '    {0} deleted..'.format(instance)
-                
+
+    def delete_depended_model_instances(self, model_cls):
+        """ Deletes all scheduled and visit models that depend on the given model"""
+#        app_name = model_cls._meta.app_label
+        for model in self.get_depended_models(model_cls):
+            self.reset_model(model)
+
+    def get_depended_models(self, fk_cls):
+        """Returns a list of scheduled and visit models that depend 
+        (ForeignKey,OneToOneField) of the given model
+        """
+        _models = []
+#        if not app_name:
+#            raise TypeError('Parameter app_name cannot be None.')
+#        app = get_app(app_name)
+#        models = get_models(app)
+        membershipform_models = self.helper.get_membershipform_models()
+        scheduled_models = self.helper.get_scheduled_models()
+        form_models = membershipform_models + scheduled_models
+        '''For each visit and membership model check if it has a field that is a 
+        foreignkey or oneToone to the given model
+         '''
+        for model_cls in form_models:
+            for field in model_cls._meta.fields:
+                if  isinstance(field, (ForeignKey, OneToOneField)):
+                    if field.rel.to == fk_cls:
+                        if model_cls not in _models:
+                            _models.append(model_cls) 
+        return _models                       
+
     def delete_audit_instances(self, model_cls): 
-        from django.db import connection, transaction
-        cursor = connection[self.get_using_destination()].cursor()
-        print "    deleting {0} audit logs".format(model_cls.history.count()) 
+        using = self.get_using_destination()
+        print "    deleting {0} {1} audit logs".format(model_cls.history.using(using).all().count(), model_cls._meta.object_name) 
+        model_cls.history.using(using).all().delete()
+
+    def reset_model_raw_sql(self, model_cls): 
+        from django.db import connections, transaction
+        cursor = connections[self.get_using_destination()].cursor()
+        table_name=format(model_cls._meta.db_table)
+        raw_sql_query = "DELETE FROM `{0}`".format(table_name)
+        cursor.execute(raw_sql_query)
+        transaction.commit_unless_managed(using=self.get_using_destination())
+        print " deleted"
+                        
+    def delete_audit_instances_raw_sql(self, model_cls): 
+        from django.db import connections, transaction
+        cursor = connections[self.get_using_destination()].cursor()
+        print "    deleting {0} audit logs".format(model_cls.history.using(self.get_using_destination()).all().count()) 
         audit_table_name = "{0}_audit".format(model_cls._meta.db_table)
-        query = "TRUNCATE TABLE `{0}`".format(audit_table_name)
-        cursor.execute(query)
+        raw_sql_query = "DELETE FROM `{0}`".format(audit_table_name)
+        cursor.execute(raw_sql_query)
         transaction.commit_unless_managed(using=self.get_using_destination())
         print " deleted"   
             
+    def reset_listed_models(self, models):
+        for app_name, model_name in models:
+            model_cls = get_model(app_name, model_name)
+            self.reset_model(model_cls)
+
     def reset_app_models(self, app_name):
         print '    deleting for app {0}...'.format(app_name)
         for model_cls in get_models(get_app(app_name)):
